+++ CHANGELOG.md
@@ -1,8 +1,22 @@
 # Changelog
 
-
 This file contains a summary of the changes that went into each release.
 
+
+## [0.1.1] - 2018-10-22
+- Fix for headless build not running on some machines (Mominon)
+- Changed boot behaviour. Now always read boot.cfg (previously named game.cfg) unless -noboot passed.
+- Added documentation about small tools in editor
+- Fix for exec command giving scary sounding warnings
+- Fix for project using Perforce by default
+- Tweak to animation for slightly smoother 3rd person
+- Fixes to SourceCode doc (Badger0101)
+- Changed Tick error message to just be info
+- Type fixes (jfmc)
+- Updated matchmaker code 
+- Added information about contribution
+- Improved documentation about animation
+- Fix for linux version of headless server not working with redict of in-/output
+
 ## [0.1] - 2018-10-22
-### Added
 - First public release, Unite L.A. 2018